requirements.txt
# Data
pydantic==2.10.6
ulid-py==1.1.0

# Model
openai==1.77.0

# Log
loguru==0.7.3
rich>=10.14.0,<14

# Devtools
black
ipython
isort
pre-commit
pytest
pytest-sugar
pytest-xdist

# Env
python-dotenv>=0.19.0
